Add a unit test
